Upgraded to the most recent version of Valgrind from SVN (~5PM EDT on 2005-04-25) and tried to fix the following few problems:
(it doesn't crash now, which is good, but I still haven't confirmed
correctness)

1.) Floating point stack problem (as evident in FloatTest)

The problem was that PUTI (Indexed PUT) is only used to write to the
floating-point stack, but Vex currently only allows 8-byte and 1-byte
writes for PUTI, 8-byte for doubles and 1-byte for the floating-point
'tags' (not the same as our tags).  We need to make an indexed 4-byte
write, so we simply relaxed a few assertions an added a case in the
x86->IR translator to handle 4-byte PUTI's.

2.) Tags and V-bits clobbering each other because they both refer to
the same place in the shadow guest state

This change was a bit more involved.  The fundamental problem is that
Vex and Valgrind assume that tools will only need one fixed-sized
shadow guest state (for holding metadata about register,
floating-point stack, etc...).  We believe that this design decision
was made in order to make Memcheck work because Memcheck requires only
one shadow guest state for the V-bits.  However, we want to maintain
our own shadow guest state, which requires 4 bytes for EVERY field
of VexGuestX86State (we're just focusing on x86 for now).

3.) Multiplexor problem in IntTest This was dumb.  I need to simply
multiplex the values of the V-bits and pass them along within
expr2tags_Mux0X_DC.  I didn't do anything to handle this case before
so the IR was invalid.