Clean calls are intuitively the correct thing for tag MERGES (and they're faster), although they may be optimized away in some special cases.
Do a clean call for most merges and do a few dirty calls to 'anchor'
the computation so that the optimizer doesn't optimize stuff away.

For the MUX and conditional exit (i.e., if (expr1) goto
literal-value), add a no-op dirty call which takes 1 tag in and does
nothing so that the optimizer doesn't destroy it.

e.g.

Mux: 'Mux0X(exprC, expr0, exprX)'

Conceptually, the return value of processing the tags for this
expression is just a parallel Mux with the SAME condition but the tags
of the expressions:

Mux0X(exprC, tag(expr0), tag(exprX))

Although exprC is an IRAtom (temp or constant), a super-complicated
expression might have resulted in that temp.  What if exprC is TEMP
which is the result of evaluating a complicated expression with lots
of interesting interactions?  The IR optimizer might blow away the
tree branch containing all of the interesting tag merges if the
resulting temp is not used anywhere (since all the tag merges are
clean calls).  Thus, we need to anchor it by adding a temporary t with
a dirty call (which cannot be optimized away) such that:

t = dirty_NOP(tag(exprC))

We don't really use t for anything, but we need it as an ANCHOR to
prevent the optimizer from blowing away tag merge interactions arising
from exprC.

Conditional exits: 'if (expr1) goto literal-value'

Ditto for expr1 here.