// By some informal observation, it seems like '>' and '>=' // comparisons are translated into clean C calls.  The correct // behavior is to merge the tags of all arguments but return a new tag // of 0 so that the tags do not propagate to the results.  Without // further testing on what other opeations are translated into IR as // clean C calls, I will simply return a tag of 0 for now.
It seems like < and <= are regular comparisons but for some reason, >
and >= are handled by clean helper calls, thus leading in an asymmetry
of tag handling for comparisons.

Solution: In handleCCall_DC(), always return a tag of 0.  It seems
like the only call to this is for '<' and '<=' so far.  Revise this
later if necessary.


Also, moved some definitions around in header files.