Fixed a bug where DynComp was failing on C++ programs because it was attempting to print out comparability information for OBJECT and CLASS program points without the framework for collecting those numbers in the first place.
The solution that I'm going to adopt right now is to not print out
comparability numbers for OBJECT and CLASS program points in the .decls file
because we don't actually explicitly print out those ppt's in the .dtrace
anyways so we have no way of collecting the comparability data at runtime.

However, we CAN collect the comparability information from the execution
of normal member functions, and if Daikon is smart enough, it can properly
infer the comparability info. of OBJECT/CLASS ppts from those of member
functions.