major performance improvement regarding tags of literals:
Important optimization: We should only create a new tag
for a literal
at translation time, NOT at run-time.  Thus, every single
static instance of a literal in some place in the code
should always have the same tag associated with it, not
a unique tag for every time that line is executed.

Remember to do something special to the uf_objects
corresponding to these tags so that they never get garbage
collected (or else we may get incorrect results) -
Right now we saturate the ref_count field