- changed prog_pts_tree to hold UInt start PC addresses of functions to trace (when using the --ppt-list-file option) instead of the demangled Daikon function names.  This enables us to filter out functions that we don't want to trace at translation-time (within handle_possible_entry and handle_possible_exit in kvasir_main.c), thus gaining a lot of speed, than at run-time.
- changed the format of the program point file (generated using
--dump-ppt-file) slightly to include an UInt representation of the
function's start PC address (which is used to uniquely identify functions)
followed by the function's Daikon name, which is only used for human
comprehension.  Thus, every line looks something like:

134518900 ..main()

The one downside of this hack is that if you modify a program and
re-compile it, the function names may still be the same but the start PC
addresses may be different, so your old ppt list file may not work
anymore.