As reported by Brian Demsky in the context of Freeciv, Kvasir sometimes doesn't correctly infer the sizes of dynamically-allocated arrays of structs when probing the heap in probeAheadDiscoverHeapArraySize().  This happens when the structs are partially-initialized (some members are initialized while others are not).  The two-step approach that Kvasir currently uses to determine array sizes starts at an address and probes forward until it reaches the first set of bytes whose A-bits are unset (unallocated).  This finds the size of a contiguous region of allocated memory.  Then it goes BACKWARDS from the end until it finds the first set of bytes whose V-bits are ALL SET.  The reason for doing this is so that Kvasir only outputs the first n initialized elements of an array rather than all elements.  This is often useful.  For example, if someone allocates a 1000-element int buffer but only fills up the first few elements, it's a waste of space to print out all 1000 elements because most of them are uninitialized.
However, this going backwards optimization turned out to be overly hasty because
it does not properly account for structs that are PARTIALLY-INITIALIZED.
Because it requires ALL bytes of an element to have V-bits set, then a
partially-initialized struct will count as uninitialized.

The easy and effective fix for this is to relax the initialization condition to
say that an element is initialized when AT LEAST one byte within it has its
V-bits set.  This will properly handle partially initialized structs.