When testing on CTAS, I determined that one function, determineTypedefNameForEntry() accounted for a HUGE performance slowdown during initialization.  In order to speed this up, what I did was to add a hashtable in typedata.c named typedef_names_map:
Key: (unsigned int) target_type_ID (the dwarf_entry ID that a typedef entry
points to)
Value: char* containing the NAME of the typedef entry with the specified
target_type_ID

Instead of calling determineTypedefNameForEntry, which does a O(n) array
traversal to find the name of a typedef entry whose target_type_ID matches a
given dwarf_entry's ID, we can now simply do a hashtable look-up with the
dwarf_entry's ID as the key.  When it finds an entry such that the
target_type_ID (key) matches the input, then it returns the name, which is
exactly what we need.  Presto!  We go from O(n) to O(1), which saves us LOTS of
time.