The current default behavior is to track comparability separately for the ENTRY and EXIT program points for each function.  However, this sometimes produces comparability numbers that Daikon cannot properly process, because it requires (sometimes, I'm not sure exactly when) that variables belong to the same comparability sets during the function entry/exit program points.
Thus, I propose to set the default behavior to keep only ONE set of data
structures to track variable comparability for each program point (the EXIT set
of data structures because it encompasses all of the variables at ENTRY plus the
return value derived variables), so that if two variables are comparable at
either ENTRY or EXIT, that merging will be reflected in the single set of data
structures.  This is guaranteed to produce the same comparability numbers for
variables at function entry/exit, which is what Daikon likes.

However, for more accurate results, I have included a --separate-entry-exit-comp
command-line option that allows for the previous behavior --- separate tracking
for entry/exit.  This gives more accuracy but may produce output that Daikon
won't accept.