added an is_declaration field to collection_type (struct/union) members of dwarf_entry_array in order to allow the DWARF parser to distinguish between empty declarations of structs/unions in compilation units which are only used for their names for linking purposes and the actual REAL declaration with the actual content.
Added code to make sure that every variable that belongs to a struct/union
type is linked properly to an actual REAL entry which lists its members
and not a fake declaration entry.

This fixes a limitation in Kvasir where it sometimes cannot traverse
inside of structs because some variables are linked to an empty declaration
entry of that struct's name in the DWARF debugging info.