Officially made my new improved variable visiting code the default option (it can still be turned of with the USE_EXP_VISIT_CODE macro in decls-output.h).  This gives significantly better results when traversing through arrays of pointers and other constructs often found in real programs, much better than the old version which assumed that everything was nested arrays inlined within structures and such.
The fundamental difference between the new code (implemented as
visitSingleVar(), visitSequence(), and friends) and the old code
(implemented as the super-ugly outputDaikonVar()) is that the new code
treats sequences as an array of pointers whereas the old code treats
sequences as simply an array with a certain size for each element.
The 'array-of-pointers' approach works much better because each element
can be dereferenced individually.

Hopefully this will continue to get more stable as Kvasir development
progresses, and soon I will be able to erase the old code entirely from
the source files.